requirements.txt
alembic==1.11.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
dnspython==2.8.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
gunicorn==21.2.0
html5lib==1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pillow==12.0.0
psycopg==3.2.11
psycopg-binary==3.2.11
pycparser==2.23
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyMySQL==1.1.1
pypdf==6.2.0
python-bidi==0.6.7
python-dotenv==1.0.1
PyYAML==6.0.3
reportlab==4.0.9
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.39.0
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.1.2
xhtml2pdf==0.2.15
